GraphProperties

Documentation for GraphProperties.

GraphProperties.Arboricity

GraphProperties.AssortativityCoefficient

GraphProperties.BipartiteDimension

GraphProperties.Boxicity

GraphProperties.CheegerConstant

GraphProperties.Choosability

GraphProperties.ChromaticIndex

GraphProperties.ChromaticNumber

GraphProperties.Circumference

GraphProperties.CliqueNumber

GraphProperties.CliqueWidth

GraphProperties.CycleRank

GraphProperties.Degeneracy

GraphProperties.DegreeSequence

GraphProperties.Diameter

GraphProperties.DigraphIsDAG

GraphProperties.DigraphIsOrientation

GraphProperties.DigraphIsPlanar

GraphProperties.DigraphIsStronglyConnected

GraphProperties.DigraphIsWeaklyConnected

GraphProperties.DominationNumber

GraphProperties.EdgeConnectivity

GraphProperties.EdgeCoverNumber

GraphProperties.FeedbackVertexSetNumber

GraphProperties.FractionalChromaticNumber

GraphProperties.FractionalMatchingNumber

GraphProperties.Girth

GraphProperties.GlobalClusteringCoefficient

GraphProperties.GraphIsBipartite

GraphProperties.GraphIsCartesianProduct

GraphProperties.GraphIsChordal

GraphProperties.GraphIsChordalBipartite

GraphProperties.GraphIsCircleGraph

GraphProperties.GraphIsCograph

GraphProperties.GraphIsComparabilityGraph

GraphProperties.GraphIsComplete

GraphProperties.GraphIsCompleteBipartite

GraphProperties.GraphIsConnected

GraphProperties.GraphIsCycle

GraphProperties.GraphIsDistanceHereditary

GraphProperties.GraphIsDistanceRegular

GraphProperties.GraphIsDistanceTransitive

GraphProperties.GraphIsDuallyChordal

GraphProperties.GraphIsForest

GraphProperties.GraphIsIndifferenceGraph

GraphProperties.GraphIsIntervalGraph

GraphProperties.GraphIsMaximalOuterplanar

GraphProperties.GraphIsMedianGraph

GraphProperties.GraphIsMeynielGraph

GraphProperties.GraphIsModular

GraphProperties.GraphIsOuterplanar

GraphProperties.GraphIsPath

GraphProperties.GraphIsPerfect

GraphProperties.GraphIsPermutationGraph

GraphProperties.GraphIsPlanar

GraphProperties.GraphIsPtolemaic

GraphProperties.GraphIsRegular

GraphProperties.GraphIsSplitGraph

GraphProperties.GraphIsSquaregraph

GraphProperties.GraphIsStar

GraphProperties.GraphIsStronglyChordal

GraphProperties.GraphIsStronglyRegular

GraphProperties.GraphIsTree

GraphProperties.GraphIsTriangleFree

GraphProperties.GraphIsTriviallyPerfect

GraphProperties.GraphProperty

GraphProperties.HadwigerNumber

GraphProperties.IndependenceNumber

GraphProperties.IntersectionNumber

GraphProperties.IsDirectedGraph

GraphProperties.IsUndirectedGraph

GraphProperties.MatchingNumber

GraphProperties.MaximumDegree

GraphProperties.MaximumIndegree

GraphProperties.MaximumOutdegree

GraphProperties.MinimumDegree

GraphProperties.MinimumIndegree

GraphProperties.MinimumOutdegree

GraphProperties.NumberOfArcs

GraphProperties.NumberOfConnectedComponents

GraphProperties.NumberOfEdges

GraphProperties.NumberOfStronglyConnectedComponents

GraphProperties.NumberOfVertices

GraphProperties.NumberOfWeaklyConnectedComponents

GraphProperties.Pathwidth

GraphProperties.PropertyComparison

GraphProperties.Radius

GraphProperties.Sphericity

GraphProperties.Splittance

GraphProperties.Strength

GraphProperties.StrongDominationNumber

GraphProperties.TreeDepth

GraphProperties.Treewidth

GraphProperties.Triameter

GraphProperties.TwinWidth

GraphProperties.VertexConnectivity

GraphProperties.VertexCoverNumber

GraphProperties.graph_property_type

GraphProperties.GraphProperty — Type

GraphProperty::Type

Abstract supertype for graph properties.

The only type parameter is the type of each value of the property.

GraphProperties.graph_property_type — Function

graph_property_type(::GraphProperty{T}) -> T

Get the type of each value of the property.

It is not allowed to add methods to this function.

GraphProperties.PropertyComparison — Type

PropertyComparison(comparison, property, value)

Turn a GraphProperty{<:Real} into a GraphProperty{Bool} by attaching a comparison against a value.

GraphProperties.Arboricity — Type

Arboricity()

The arboricity of an undirected graph.

Example:

julia> using GraphProperties

julia> Arboricity() isa GraphProperty{Integer}
true

GraphProperties.AssortativityCoefficient — Type

AssortativityCoefficient()

The assortativity coefficient of an undirected graph.

Example:

julia> using GraphProperties

julia> AssortativityCoefficient() isa GraphProperty{Real}
true

GraphProperties.BipartiteDimension — Type

BipartiteDimension()

The bipartite dimension of an undirected graph. Also known as:

biclique cover number

Example:

julia> using GraphProperties

julia> BipartiteDimension() isa GraphProperty{Integer}
true

GraphProperties.Boxicity — Type

Boxicity()

The boxicity of an undirected graph.

Example:

julia> using GraphProperties

julia> Boxicity() isa GraphProperty{Integer}
true

GraphProperties.CheegerConstant — Type

CheegerConstant()

The Cheeger constant of an undirected graph. Also known as:

Cheeger number

isoperimetric number

Example:

julia> using GraphProperties

julia> CheegerConstant() isa GraphProperty{Real}
true

GraphProperties.Choosability — Type

Choosability()

The choosability of an undirected graph. Also known as:

list colorability

list chromatic number

Example:

julia> using GraphProperties

julia> Choosability() isa GraphProperty{Integer}
true

GraphProperties.ChromaticIndex — Type

ChromaticIndex()

The chromatic index of an undirected graph. Also known as:

edge chromatic number

Example:

julia> using GraphProperties

julia> ChromaticIndex() isa GraphProperty{Integer}
true

GraphProperties.ChromaticNumber — Type

ChromaticNumber()

The chromatic number of an undirected graph.

Example:

julia> using GraphProperties

julia> ChromaticNumber() isa GraphProperty{Integer}
true

GraphProperties.Circumference — Type

Circumference()

The circumference of an undirected graph.

Example:

julia> using GraphProperties

julia> Circumference() isa GraphProperty{Real}
true

GraphProperties.CliqueNumber — Type

CliqueNumber()

The clique number of an undirected graph.

Example:

julia> using GraphProperties

julia> CliqueNumber() isa GraphProperty{Integer}
true

GraphProperties.CliqueWidth — Type

CliqueWidth()

The clique-width of an undirected graph.

Example:

julia> using GraphProperties

julia> CliqueWidth() isa GraphProperty{Integer}
true

GraphProperties.CycleRank — Type

CycleRank()

The cycle rank of a directed graph.

Example:

julia> using GraphProperties

julia> CycleRank() isa GraphProperty{Integer}
true

GraphProperties.Degeneracy — Type

Degeneracy()

The degeneracy of an undirected graph. Also known as:

width

linkage

Example:

julia> using GraphProperties

julia> Degeneracy() isa GraphProperty{Integer}
true

GraphProperties.DegreeSequence — Type

DegreeSequence()

The degree sequence of an undirected graph.

Example:

julia> using GraphProperties

julia> DegreeSequence() isa GraphProperty{AbstractVector{<:Integer}}
true

GraphProperties.Diameter — Type

Diameter()

The diameter of a connected undirected graph.

Example:

julia> using GraphProperties

julia> Diameter() isa GraphProperty{Integer}
true

GraphProperties.DigraphIsDAG — Type

DigraphIsDAG()

Is a directed graph a directed acyclic graph? Also known as:

a DAG

Example:

julia> using GraphProperties

julia> DigraphIsDAG() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsOrientation — Type

DigraphIsOrientation()

Is a directed graph an orientation?

Example:

julia> using GraphProperties

julia> DigraphIsOrientation() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsPlanar — Type

DigraphIsPlanar()

Is a directed graph planar?

Example:

julia> using GraphProperties

julia> DigraphIsPlanar() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsStronglyConnected — Type

DigraphIsStronglyConnected()

Is a directed graph strongly connected?

Example:

julia> using GraphProperties

julia> DigraphIsStronglyConnected() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsWeaklyConnected — Type

DigraphIsWeaklyConnected()

Is a directed graph weakly connected?

Example:

julia> using GraphProperties

julia> DigraphIsWeaklyConnected() isa GraphProperty{Bool}
true

GraphProperties.DominationNumber — Type

DominationNumber()

The domination number of an undirected graph.

Example:

julia> using GraphProperties

julia> DominationNumber() isa GraphProperty{Integer}
true

GraphProperties.EdgeConnectivity — Type

EdgeConnectivity()

The edge-connectivity of an undirected graph.

Example:

julia> using GraphProperties

julia> EdgeConnectivity() isa GraphProperty{Integer}
true

GraphProperties.EdgeCoverNumber — Type

EdgeCoverNumber()

The edge cover number of an undirected graph.

Example:

julia> using GraphProperties

julia> EdgeCoverNumber() isa GraphProperty{Integer}
true

GraphProperties.FeedbackVertexSetNumber — Type

FeedbackVertexSetNumber()

The feedback vertex set number of a graph.

Example:

julia> using GraphProperties

julia> FeedbackVertexSetNumber() isa GraphProperty{Integer}
true

GraphProperties.FractionalChromaticNumber — Type

FractionalChromaticNumber()

The fractional chromatic number of an undirected graph.

Example:

julia> using GraphProperties

julia> FractionalChromaticNumber() isa GraphProperty{Real}
true

GraphProperties.FractionalMatchingNumber — Type

FractionalMatchingNumber()

The fractional matching number of an undirected graph.

Example:

julia> using GraphProperties

julia> FractionalMatchingNumber() isa GraphProperty{Real}
true

GraphProperties.Girth — Type

Girth()

The girth of an undirected graph.

Example:

julia> using GraphProperties

julia> Girth() isa GraphProperty{Real}
true

GraphProperties.GlobalClusteringCoefficient — Type

GlobalClusteringCoefficient()

The global clustering coefficient of an undirected graph.

Example:

julia> using GraphProperties

julia> GlobalClusteringCoefficient() isa GraphProperty{Real}
true

GraphProperties.GraphIsBipartite — Type

GraphIsBipartite()

Is an undirected graph bipartite?

Example:

julia> using GraphProperties

julia> GraphIsBipartite() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCartesianProduct — Type

GraphIsCartesianProduct()

Is an undirected graph a Cartesian product of two graphs? Also known as:

a box product

Example:

julia> using GraphProperties

julia> GraphIsCartesianProduct() isa GraphProperty{Bool}
true

GraphProperties.GraphIsChordal — Type

GraphIsChordal()

Is an undirected graph chordal?

Example:

julia> using GraphProperties

julia> GraphIsChordal() isa GraphProperty{Bool}
true

GraphProperties.GraphIsChordalBipartite — Type

GraphIsChordalBipartite()

Is an undirected graph chordal bipartite?

Example:

julia> using GraphProperties

julia> GraphIsChordalBipartite() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCircleGraph — Type

GraphIsCircleGraph()

Is an undirected graph a circle graph?

Example:

julia> using GraphProperties

julia> GraphIsCircleGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCograph — Type

GraphIsCograph()

Is an undirected graph a cograph?

Example:

julia> using GraphProperties

julia> GraphIsCograph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsComparabilityGraph — Type

GraphIsComparabilityGraph()

Is an undirected graph a comparability graph?

Example:

julia> using GraphProperties

julia> GraphIsComparabilityGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsComplete — Type

GraphIsComplete()

Is an undirected graph complete?

Example:

julia> using GraphProperties

julia> GraphIsComplete() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCompleteBipartite — Type

GraphIsCompleteBipartite()

Is an undirected graph complete bipartite? Also known as:

a biclique

Example:

julia> using GraphProperties

julia> GraphIsCompleteBipartite() isa GraphProperty{Bool}
true

GraphProperties.GraphIsConnected — Type

GraphIsConnected()

Is an undirected graph connected?

Example:

julia> using GraphProperties

julia> GraphIsConnected() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCycle — Type

GraphIsCycle()

Is an undirected graph a cycle?

Example:

julia> using GraphProperties

julia> GraphIsCycle() isa GraphProperty{Bool}
true

GraphProperties.GraphIsDistanceHereditary — Type

GraphIsDistanceHereditary()

Is an undirected graph distance-hereditary?

Example:

julia> using GraphProperties

julia> GraphIsDistanceHereditary() isa GraphProperty{Bool}
true

GraphProperties.GraphIsDistanceRegular — Type

GraphIsDistanceRegular()

Is an undirected graph distance-regular?

Example:

julia> using GraphProperties

julia> GraphIsDistanceRegular() isa GraphProperty{Bool}
true

GraphProperties.GraphIsDistanceTransitive — Type

GraphIsDistanceTransitive()

Is an undirected graph distance-transitive?

Example:

julia> using GraphProperties

julia> GraphIsDistanceTransitive() isa GraphProperty{Bool}
true

GraphProperties.GraphIsDuallyChordal — Type

GraphIsDuallyChordal()

Is an undirected graph dually chordal?

Example:

julia> using GraphProperties

julia> GraphIsDuallyChordal() isa GraphProperty{Bool}
true

GraphProperties.GraphIsForest — Type

GraphIsForest()

Is an undirected graph a forest?

Example:

julia> using GraphProperties

julia> GraphIsForest() isa GraphProperty{Bool}
true

GraphProperties.GraphIsIndifferenceGraph — Type

GraphIsIndifferenceGraph()

Is an undirected graph an indifference graph?

Example:

julia> using GraphProperties

julia> GraphIsIndifferenceGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsIntervalGraph — Type

GraphIsIntervalGraph()

Is an undirected graph an interval graph?

Example:

julia> using GraphProperties

julia> GraphIsIntervalGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsMaximalOuterplanar — Type

GraphIsMaximalOuterplanar()

Is an undirected graph maximal outerplanar?

Example:

julia> using GraphProperties

julia> GraphIsMaximalOuterplanar() isa GraphProperty{Bool}
true

GraphProperties.GraphIsMedianGraph — Type

GraphIsMedianGraph()

Is an undirected graph a median graph?

Example:

julia> using GraphProperties

julia> GraphIsMedianGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsMeynielGraph — Type

GraphIsMeynielGraph()

Is an undirected graph a Meyniel graph?

Example:

julia> using GraphProperties

julia> GraphIsMeynielGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsModular — Type

GraphIsModular()

Is an undirected graph modular?

Example:

julia> using GraphProperties

julia> GraphIsModular() isa GraphProperty{Bool}
true

GraphProperties.GraphIsOuterplanar — Type

GraphIsOuterplanar()

Is an undirected graph outerplanar?

Example:

julia> using GraphProperties

julia> GraphIsOuterplanar() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPath — Type

GraphIsPath()

Is an undirected graph a path?

Example:

julia> using GraphProperties

julia> GraphIsPath() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPerfect — Type

GraphIsPerfect()

Is an undirected graph perfect?

Example:

julia> using GraphProperties

julia> GraphIsPerfect() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPermutationGraph — Type

GraphIsPermutationGraph()

Is an undirected graph a permutation graph?

Example:

julia> using GraphProperties

julia> GraphIsPermutationGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPlanar — Type

GraphIsPlanar()

Is an undirected graph planar?

Example:

julia> using GraphProperties

julia> GraphIsPlanar() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPtolemaic — Type

GraphIsPtolemaic()

Is an undirected graph Ptolemaic?

Example:

julia> using GraphProperties

julia> GraphIsPtolemaic() isa GraphProperty{Bool}
true

GraphProperties.GraphIsRegular — Type

GraphIsRegular()

Is an undirected graph regular?

Example:

julia> using GraphProperties

julia> GraphIsRegular() isa GraphProperty{Bool}
true

GraphProperties.GraphIsSplitGraph — Type

GraphIsSplitGraph()

Is an undirected graph a split graph?

Example:

julia> using GraphProperties

julia> GraphIsSplitGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsSquaregraph — Type

GraphIsSquaregraph()

Is an undirected graph a squaregraph?

Example:

julia> using GraphProperties

julia> GraphIsSquaregraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsStar — Type

GraphIsStar()

Is an undirected graph a star?

Example:

julia> using GraphProperties

julia> GraphIsStar() isa GraphProperty{Bool}
true

GraphProperties.GraphIsStronglyChordal — Type

GraphIsStronglyChordal()

Is an undirected graph strongly chordal?

Example:

julia> using GraphProperties

julia> GraphIsStronglyChordal() isa GraphProperty{Bool}
true

GraphProperties.GraphIsStronglyRegular — Type

GraphIsStronglyRegular()

Is an undirected graph strongly regular?

Example:

julia> using GraphProperties

julia> GraphIsStronglyRegular() isa GraphProperty{Bool}
true

GraphProperties.GraphIsTree — Type

GraphIsTree()

Is an undirected graph a tree?

Example:

julia> using GraphProperties

julia> GraphIsTree() isa GraphProperty{Bool}
true

GraphProperties.GraphIsTriangleFree — Type

GraphIsTriangleFree()

Is an undirected graph triangle-free?

Example:

julia> using GraphProperties

julia> GraphIsTriangleFree() isa GraphProperty{Bool}
true

GraphProperties.GraphIsTriviallyPerfect — Type

GraphIsTriviallyPerfect()

Is an undirected graph trivially perfect?

Example:

julia> using GraphProperties

julia> GraphIsTriviallyPerfect() isa GraphProperty{Bool}
true

GraphProperties.HadwigerNumber — Type

HadwigerNumber()

The Hadwiger number of an undirected graph. Also known as:

contraction clique number

homomorphism degree

Example:

julia> using GraphProperties

julia> HadwigerNumber() isa GraphProperty{Integer}
true

GraphProperties.IndependenceNumber — Type

IndependenceNumber()

The independence number of an undirected graph.

Example:

julia> using GraphProperties

julia> IndependenceNumber() isa GraphProperty{Integer}
true

GraphProperties.IntersectionNumber — Type

IntersectionNumber()

The intersection number of an undirected graph. Also known as:

R-content

edge clique cover number

clique cover number

Example:

julia> using GraphProperties

julia> IntersectionNumber() isa GraphProperty{Integer}
true

GraphProperties.IsDirectedGraph — Type

IsDirectedGraph()

Is a graph a directed graph? Also known as:

a digraph

Example:

julia> using GraphProperties

julia> IsDirectedGraph() isa GraphProperty{Bool}
true

GraphProperties.IsUndirectedGraph — Type

IsUndirectedGraph()

Is a graph an undirected graph?

Example:

julia> using GraphProperties

julia> IsUndirectedGraph() isa GraphProperty{Bool}
true

GraphProperties.MatchingNumber — Type

MatchingNumber()

The matching number of an undirected graph.

Example:

julia> using GraphProperties

julia> MatchingNumber() isa GraphProperty{Integer}
true

GraphProperties.MaximumDegree — Type

MaximumDegree()

The maximum degree of an undirected graph.

Example:

julia> using GraphProperties

julia> MaximumDegree() isa GraphProperty{Integer}
true

GraphProperties.MaximumIndegree — Type

MaximumIndegree()

The maximum indegree of a directed graph.

Example:

julia> using GraphProperties

julia> MaximumIndegree() isa GraphProperty{Integer}
true

GraphProperties.MaximumOutdegree — Type

MaximumOutdegree()

The maximum outdegree of a directed graph.

Example:

julia> using GraphProperties

julia> MaximumOutdegree() isa GraphProperty{Integer}
true

GraphProperties.MinimumDegree — Type

MinimumDegree()

The minimum degree of an undirected graph.

Example:

julia> using GraphProperties

julia> MinimumDegree() isa GraphProperty{Integer}
true

GraphProperties.MinimumIndegree — Type

MinimumIndegree()

The minimum indegree of a directed graph.

Example:

julia> using GraphProperties

julia> MinimumIndegree() isa GraphProperty{Integer}
true

GraphProperties.MinimumOutdegree — Type

MinimumOutdegree()

The minimum outdegree of a directed graph.

Example:

julia> using GraphProperties

julia> MinimumOutdegree() isa GraphProperty{Integer}
true

GraphProperties.NumberOfArcs — Type

NumberOfArcs()

The number of arcs of a directed graph.

Example:

julia> using GraphProperties

julia> NumberOfArcs() isa GraphProperty{Integer}
true

GraphProperties.NumberOfConnectedComponents — Type

NumberOfConnectedComponents()

The number of connected components of an undirected graph.

Example:

julia> using GraphProperties

julia> NumberOfConnectedComponents() isa GraphProperty{Integer}
true

GraphProperties.NumberOfEdges — Type

NumberOfEdges()

The number of edges of an undirected graph. Also known as:

size

Example:

julia> using GraphProperties

julia> NumberOfEdges() isa GraphProperty{Integer}
true

GraphProperties.NumberOfStronglyConnectedComponents — Type

NumberOfStronglyConnectedComponents()

The number of strongly connected components of a directed graph.

Example:

julia> using GraphProperties

julia> NumberOfStronglyConnectedComponents() isa GraphProperty{Integer}
true

GraphProperties.NumberOfVertices — Type

NumberOfVertices()

The number of vertices of a graph. Also known as:

order

Example:

julia> using GraphProperties

julia> NumberOfVertices() isa GraphProperty{Integer}
true

GraphProperties.NumberOfWeaklyConnectedComponents — Type

NumberOfWeaklyConnectedComponents()

The number of weakly connected components of a directed graph.

Example:

julia> using GraphProperties

julia> NumberOfWeaklyConnectedComponents() isa GraphProperty{Integer}
true

GraphProperties.Pathwidth — Type

Pathwidth()

The pathwidth of an undirected graph. Also known as:

interval thickness

vertex separation number

node searching number

Example:

julia> using GraphProperties

julia> Pathwidth() isa GraphProperty{Integer}
true

GraphProperties.Radius — Type

Radius()

The radius of a connected undirected graph.

Example:

julia> using GraphProperties

julia> Radius() isa GraphProperty{Integer}
true

GraphProperties.Sphericity — Type

Sphericity()

The sphericity of an undirected graph.

Example:

julia> using GraphProperties

julia> Sphericity() isa GraphProperty{Integer}
true

GraphProperties.Splittance — Type

Splittance()

The splittance of an undirected graph.

Example:

julia> using GraphProperties

julia> Splittance() isa GraphProperty{Integer}
true

GraphProperties.Strength — Type

Strength()

The strength of an undirected graph.

Example:

julia> using GraphProperties

julia> Strength() isa GraphProperty{Real}
true

GraphProperties.StrongDominationNumber — Type

StrongDominationNumber()

The strong domination number of an undirected graph.

Example:

julia> using GraphProperties

julia> StrongDominationNumber() isa GraphProperty{Integer}
true

GraphProperties.TreeDepth — Type

TreeDepth()

The tree-depth of an undirected graph.

Example:

julia> using GraphProperties

julia> TreeDepth() isa GraphProperty{Integer}
true

GraphProperties.Treewidth — Type

Treewidth()

The treewidth of an undirected graph.

Example:

julia> using GraphProperties

julia> Treewidth() isa GraphProperty{Integer}
true

GraphProperties.Triameter — Type

Triameter()

The triameter of a connected undirected graph.

Example:

julia> using GraphProperties

julia> Triameter() isa GraphProperty{Integer}
true

GraphProperties.TwinWidth — Type

TwinWidth()

The twin-width of an undirected graph.

Example:

julia> using GraphProperties

julia> TwinWidth() isa GraphProperty{Integer}
true

GraphProperties.VertexConnectivity — Type

VertexConnectivity()

The vertex-connectivity of an undirected graph. Also known as:

connectivity

Example:

julia> using GraphProperties

julia> VertexConnectivity() isa GraphProperty{Integer}
true

GraphProperties.VertexCoverNumber — Type

VertexCoverNumber()

The vertex cover number of an undirected graph.

Example:

julia> using GraphProperties

julia> VertexCoverNumber() isa GraphProperty{Integer}
true