Basics

This module contains basic graph properties functions.

GraphProperties.Basics.diameter — Method

diameter(g::AbstractGraph{T}) where T <: Integer

Return the diameter of g.

Arguments

g::AbstractGraph{T}: A graph.

Returns

The diameter of g.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = PathGraph(5);

julia> diameter(g)
4

GraphProperties.Basics.girth — Method

girth(g::SimpleGraph{T}) where T <: Integer

Return the girth of g.

The girth of a graph is the length of the shortest cycle in the graph.

Arguments

g::SimpleGraph{T}: A graph.

Returns

The girth of g.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = PathGraph(5);

julia> girth(g)
Inf

GraphProperties.Basics.is_bull_free — Method

is_bull_free(G::SimpleGraph{T})

Return true if no induced subgraph of G is a bull, and false otherwise.

A bull is a graph formed by adding two pendant edges to separate verticies of a triangle.

Arguments

G::SimpleGraph{T}: A graph.

Returns

true if G is bull-free, and false otherwise.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = Graphs.SimpleGraph(5);

julia> add_edge!(g, 1, 2);

julia> add_edge!(g, 2, 3);

julia> add_edge!(g, 3, 1);

julia> add_edge!(g, 2, 4);

julia> add_edge!(g, 2, 5);

julia> is_bull_free(g)
true

GraphProperties.Basics.is_claw_free — Method

is_claw_free(G::SimpleGraph{T})

Return true if no induced subgraph of G is a claw, and false otherwise.

A claw is a graph formed by adding three pendant edges to a single vertex.

Arguments

G::SimpleGraph{T}: A graph.

Returns

true if G is claw-free, and false otherwise.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = Graphs.SimpleGraph(5);

julia> add_edge!(g, 1, 2);

julia> add_edge!(g, 1, 3);

julia> add_edge!(g, 1, 4);

julia> add_edge!(g, 1, 5);

julia> is_claw_free(g)
false

GraphProperties.Basics.is_triangle_free — Method

is_triangle_free(G::SimpleGraph{T})

Return true if G is triangle-free, and false otherwise.

A graph is triangle-free if it does not contain any induced subgraph isomorphic to a triangle.

Arguments

G::SimpleGraph{T}: A graph.

Returns

true if G is triangle-free, and false otherwise.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = PathGraph(5);

julia> is_triangle_free(g)
true

GraphProperties.Basics.order — Method

order(g::AbstractGraph{T}) where T <: Integer

Return the order (the number of vertices) of g.

Arguments

g::AbstractGraph{T}: A graph.

Returns

The order of g.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = PathGraph(5);

julia> order(g)
5

GraphProperties.Basics.radius — Method

radius(g::AbstractGraph{T}) where T <: Integer

Return the radius of g.

The radius of a graph is the minimum eccentricity of its vertices. In other words, it is the minimum distance from any vertex to the farthest vertex from it.

Arguments

g::AbstractGraph{T}: A graph.

Returns

The radius of g.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = PathGraph(5);

julia> radius(g)
2

GraphProperties.Basics.size — Method

size(g::AbstractGraph{T}) where T <: Integer

Return the size (the number of edges) of g.

Arguments

g::AbstractGraph{T}: A graph.

Returns

The size of g.

Example

julia> using Graphs

julia> using GraphProperties.Basics

julia> g = CompleteGraph(5);

julia> size(g)
10