Communities

This module contains functions for community detection algorithms and related operations. Most code is derived from the Graph Communities Julia Package. Algorithms include label propagation, pagerank, k-clique, and others. The module also includes other methods for graph construction, community analysis, and visualization.

GraphProperties.Communities.ChainedCliques — Type

ChainedCliques <: CommunityGraph

A graph structure that represents a series of connected cliques.

Fields

num_cliques::Int: The number of cliques in the graph.

clique_size::Int: The number of nodes in each clique.

Examples

graph_info = ChainedCliques(num_cliques=5, clique_size=4)
graph = generate(graph_info)

GraphProperties.Communities.CommunityDetectionAlgorithm — Type

CommunityDetectionAlgorithm

Abstract type representing a community detection algorithm. This serves as the base type for various community detection algorithms implemented in the package.

Notes

All custom community detection algorithms in the package should subtype this abstract type. See Louvain, KClique, etc. for concrete implementations.

GraphProperties.Communities.CommunityGraph — Type

CommunityGraph

An abstract type representing a graph with community structure.

GraphProperties.Communities.FastLPA — Type

FastLPA <: CommunityDetectionAlgorithm

The (Fast) Label Propagation algorithm for community detection in networks.

The Label Propagation algorithm identifies communities based on the diffusion of labels throughout the graph. Nodes adopt the label that is most common among their neighbors. This process iteratively refines labels until a consensus or stable state is reached, where nodes have predominantly the same label as their neighbors.

The algorithm can be run in either synchronous or asynchronous mode:

Synchronous: All nodes update their labels simultaneously in each iteration.

Asynchronous: Nodes update their labels in a random order (not yet implemented).

Arguments

synchronous::Bool: If true, updates labels in synchronous mode; if false (default),

updates labels in asynchronous mode.

max_iter::Int: Maximum number of iterations (default is 100). If the algorithm doesn't

converge within this number of iterations, it will halt and return the current vector.

Usage

communities = compute(FastLPA(), graph)            # Synchronous (default)

References

Raghavan, U. N., Albert, R., & Kumara, S. (2007). Near linear time algorithm to detect

community structures in large-scale networks. Physical review E, 76(3), 036106.

GraphProperties.Communities.KClique — Type

KClique <: CommunityDetectionAlgorithm

The K-Clique Percolation algorithm for community detection in networks.

This method identifies communities based on the presence of K-clique (with K = 3) structures within the graph, where a K-clique is a fully connected subgraph of K nodes. Two K-cliques are adjacent if they share K-1 nodes, and a community is defined as the union of K-cliques that can be reached from each other through a series of adjacent K-cliques.

Usage

communities = compute(KClique(), graph)

References

Palla, G., Derényi, I., Farkas, I., & Vicsek, T. (2005). Uncovering the overlapping community structure of complex networks in nature and society. Nature, 435(7043), 814-818.

GraphProperties.Communities.KarateClub — Type

KarateClub <: CommunityGraph

The KarateClub graph, often referred to as the "Zachary's Karate Club", is a social network of friendships between 34 members of a karate club at a US university in the 1970s. This dataset has become a standard benchmark in community detection literature because of its well-documented community structure.

The graph captures the observed friendships between the 34 members. During the course of the study, the club split into two communities due to a conflict, making it a valuable dataset for studying community detection algorithms.

Usage

graph = generate(KarateClub())

References

Zachary, W. W. (1977). An information flow model for conflict and fission in small groups. Journal of anthropological research, 452-473.

GraphProperties.Communities.LabelPropagation — Type

LabelPropagation <: CommunityDetectionAlgorithm

The Label Propagation algorithm for community detection in networks.

The Label Propagation algorithm identifies communities based on the diffusion of labels throughout the graph. Nodes adopt the label that is most common among their neighbors. This process iteratively refines labels until a consensus or stable state is reached, where nodes have predominantly the same label as their neighbors.

The algorithm can be run in either synchronous or asynchronous mode:

Synchronous: All nodes update their labels simultaneously in each iteration.

Asynchronous: Nodes update their labels in a random order.

Arguments

synchronous::Bool: If true, updates labels in synchronous mode; if false (default),

updates labels in asynchronous mode.

max_iter::Int: Maximum number of iterations (default is 100). If the algorithm doesn't

converge within this number of iterations, it will halt and return the current vector.

Usage

communities = compute(LabelPropagation(), graph)            # Asynchronous (default)
communities = compute(LabelPropagation(sync=true), graph)   # Synchronous

References

Raghavan, U. N., Albert, R., & Kumara, S. (2007). Near linear time algorithm to detect

community structures in large-scale networks. Physical review E, 76(3), 036106.

GraphProperties.Communities.Louvain — Type

Louvain <: CommunityDetectionAlgorithm

The Louvain algorithm for community detection in networks.

This method optimizes the modularity of partitions of the graph. It follows a greedy optimization approach that generally operates in time (O(n \log n)) , making it efficient for large-scale networks.

Usage

communities = compute(Louvain(), graph)

References

Blondel, V. D., Guillaume, J. L., Lambiotte, R., & Lefebvre, E. (2008). Fast unfolding

of communities in large networks. Journal of Statistical Mechanics: Theory and Experiment, 2008(10), P10008.

GraphProperties.Communities.PageRank — Type

PageRank <: CommunityDetectionAlgorithm

PageRank is an algorithm originally designed for ranking web pages in search results. However, it can also be used more broadly in networks to determine the importance of nodes within a graph. The underlying principle is that more important nodes are likely to receive more links from other nodes.

The algorithm computes a stationary distribution of a random walk on the graph where, at each step, with probability d, the walker randomly chooses an outgoing link from its current node and with probability 1 - d, it jumps to a random node in the graph.

Arguments

d::Float64: Damping factor (default is 0.85). It represents the probability that the random walker follows an outgoing edge. Typically set between 0.85 and 0.9.

tol::Float64: Tolerance for determining convergence (default is 1e-6). The algorithm stops iterating once the change between subsequent PageRank vectors is below this value.

max_iter::Int: Maximum number of iterations (default is 100). If the algorithm doesn't converge within this number of iterations, it will halt and return the current vector.

Usage

pageranks = compute(PageRank(), graph)  # Using default parameters
pageranks = compute(PageRank(d=0.9, tol=1e-7, max_iter=150), graph)

References

Page, L., Brin, S., Motwani, R., & Winograd, T. (1999). The PageRank citation ranking: Bringing order to the web. Stanford InfoLab.

GraphProperties.Communities.PlantedPartition — Type

PlantedPartition <: CommunityGraph

The PlantedPartition model, also known as the Stochastic Block Model (SBM), is a probabilistic model commonly used for generating synthetic networks with inherent community structures. This model creates a graph by partitioning nodes into distinct communities and then adding edges between nodes based on intra-community and inter-community probabilities.

Arguments

n_communities::Int: Number of communities or blocks in the graph.

nodes_per_community::Int: Number of nodes within each community.

pintra::Float64: Probability of creating an edge between two nodes within the same community. This defines the density of intra-community edges.

pinter::Float64: Probability of creating an edge between two nodes from different communities. This defines the sparsity of inter-community edges.

Typically, pintra is set to be much larger than pinter to ensure dense intra-community connections and sparse inter-community connections, thereby creating discernible community structures.

Usage

graph1 = generate(PlantedPartition())  # Using default parameters
graph2 = generate(PlantedPartition(n_communities=5, nodes_per_community=10, pintra=0.8, pinter=0.02))

References

Holland, P. W., Laskey, K. B., & Leinhardt, S. (1983). Stochastic blockmodels: First steps. Social networks, 5(2), 109-137.

GraphProperties.Communities._adaptive_regularization_pagerank — Method

_adaptive_pagerank(
    P::SparseMatrixCSC, d::Float64, tol::Float64, max_iter::Int64)::Vector{Float64}

Helper function for compute(Pagerank, g), computes the PageRank values of the nodes in graph g using the adaptive regularization PageRank algorithm.

Arguments

P::SparseMatrixCSC: The Markov Chain associated with g.

d::Float64: The initial damping factor. 0 < d < 1.0

tol::Float64: The tolerance for convergence.

max_iter::Int64: The maximum number of iterations.

Returns

A vector of Float64 where each entry represents the PageRank value of the corresponding node in the graph.

GraphProperties.Communities._classical_pagerank — Method

_classical_pagerank(
    P::SparseMatrixCSC, d::Float64, tol::Float64, max_iter::Int64)::Vector{Float64}

Helper function for compute(Pagerank, g), computes the PageRank values of the nodes in graph g using the classical PageRank algorithm.

Arguments

P::SparseMatrixCSC: The Markov Chain associated with g.

d::Float64: The damping factor.

tol::Float64: The tolerance for convergence.

max_iter::Int64: The maximum number of iterations.

Returns

A vector of Float64 where each entry represents the PageRank value of the corresponding node in the graph.

GraphProperties.Communities._draw_communities — Method

draw_communities(g::AbstractGraph, communities::Dict{Int, Int})

Internal function to draw the graph g with nodes colored based on their community assignments.

This function maps each unique community label to a distinct color and then generates a plot where each node in the graph g is colored according to its community label.

Arguments

g::AbstractGraph: The input graph.

communities::Dict{Int, Int}: A dictionary mapping each vertex to its community label.

Returns

A plot object with nodes colored based on their community label. Each unique community is represented by a unique color.

Notes

This function is an internal helper and is typically not called directly by users. It is used by higher-level community drawing functions which prepare and pass the community mapping in the form of a dictionary.

The function assumes that every vertex in the graph g is assigned a community label in the communities dictionary. If any vertex is missing a label, an error will be raised in the calling function.

The colors are chosen to be as distinguishable as possible, but for graphs with a very large number of communities, some colors may be similar.

Example

julia> g = generate(KarateClub())

julia> communities = compute(LabelPropagation(), g)

julia> draw_communities(g, communities)

GraphProperties.Communities._iterative_regularization_pagerank — Method

_iterative_pagerank(
    P::SparseMatrixCSC, d::Float64, tol::Float64, max_iter::Int64)::Vector{Float64}

Helper function for compute(Pagerank, g), computes the PageRank values of the nodes in graph g using the iterative regularization PageRank algorithm.

Arguments

P::SparseMatrixCSC: The Markov Chain associated with g.

d::Float64: The initial damping factor, from which the regularization factor is computed. 0 < d < .5

tol::Float64: The tolerance for convergence.

max_iter::Int64: The maximum number of iterations.

Returns

A vector of Float64 where each entry represents the PageRank value of the corresponding node in the graph.

GraphProperties.Communities._label_propagation_sweep! — Method

_label_propagation_sweep!(
    pres_labels::Vector{Int},
    next_labels::Vector{Int},
    seen_labels::Vector{Int},
    seen_label_counts::Vector{Int},
    edge_list::Vector{Tuple{Int,Int}}
)::Bool

Performs a single sweep of label propagation on a graph.

This function updates the labels of vertices based on the most frequent labels among their neighbors, using a provided edge list.

Arguments

pres_labels::Vector{Int}: The current labels of the vertices.

next_labels::Vector{Int}: The vector to store the updated labels.

seen_labels::Vector{Int}: A buffer vector to store observed neighbor labels.

seen_label_counts::Vector{Int}: A vector to count occurrences of each label.

edge_list::Vector{Tuple{Int,Int}}: The list of edges in the graph.

Returns

true if the labeling has reached a stationary state (no changes), false otherwise.

Notes

Assumes the edge list is lexicographically ordered.

Edges are processed sequentially, and labels are updated based on the most frequent neighbor labels.

GraphProperties.Communities._max_argmax_in_seen — Method

_max_argmax_in_seen(
    seen_label_counts::Vector{Int},
    seen_labels::Vector{Int},
    n_seen_labels::Int
)::Int

Finds the maximum label with the highest count among seen labels.

This function is used in the label propagation algorithm to determine the most popular label among a node's neighbors.

Arguments

seen_label_counts::Vector{Int}: A vector containing counts of each label seen.

seen_labels::Vector{Int}: The labels that have been seen in the current context.

n_seen_labels::Int: The number of seen labels to consider.

Returns

The label with the highest count among the seen labels. In case of a tie, the highest label is chosen.

Notes

This function is part of the internal mechanics of label propagation and operates on a subset of labels.

GraphProperties.Communities._max_argmax_in_seen_weighted — Method

_max_argmax_in_seen_weighted(
    seen_label_weights::Vector{Float64},
    seen_labels::Vector{Int},
    n_seen_labels::Int
)::Int

Finds the maximum label with the highest accumulated weight among seen labels.

This function is used in the label propagation algorithm for weighted graphs to determine the most 'influential' label among a node's neighbors, based on the weights of the edges connecting to those neighbors.

Arguments

seen_label_weights::Vector{Float64}: A vector containing the accumulated weights for each label seen.

seen_labels::Vector{Int}: The labels that have been seen in the current context.

n_seen_labels::Int: The number of seen labels to consider.

Returns

The label with the highest accumulated weight among the seen labels. In case of a tie, the highest label is chosen.

Notes

This function is part of the internal mechanics of label propagation for weighted graphs and operates on a subset of labels.

GraphProperties.Communities.compute — Method

compute(algo::FastLPA, g::SimpleGraph)

Execute the Fast Label Propagation algorithm on a graph.

This function processes a SimpleGraph using the Fast Label Propagation algorithm to perform community detection or labeling. It first preprocesses the graph to generate an edge list and the number of vertices, then applies synchronous label propagation if enabled.

Arguments

algo::FastLabelPropagation: The Fast Label Propagation algorithm instance.

g::SimpleGraph: The graph to be processed, represented as a SimpleGraph.

Returns

If algo.synchronous is true, it returns the result of synchronous label propagation; otherwise, it returns nothing.

Notes

The graph g is first converted into an edge list and the number of vertices is determined.

This function delegates to _sync_label_propagation for the actual label propagation process.

Currently, only synchronous label propagation is implemented. If algo.synchronous is false, the function will return nothing.

Example

julia> g = generate(PlantedPartition())

julia> compute(FastLPA(), g)

GraphProperties.Communities.compute — Method

compute(algo::FastLPA, g::SimpleWeightedGraph)

Execute the Fast Label Propagation algorithm on a graph.

This function processes a SimpleWeightedGraph using the Fast Label Propagation algorithm to perform community detection or labeling. It first preprocesses the graph to generate an edge list and the number of vertices, then applies synchronous label propagation if enabled.

Arguments

algo::FastLabelPropagation: The Fast Label Propagation algorithm instance.

g::SimpleWeightedGraph: The graph to be processed, represented as a SimpleWeightedGraph.

Returns

If algo.synchronous is true, it returns the result of synchronous label propagation; otherwise, it returns nothing.

Notes

The graph g is first converted into an edge list and the number of vertices is determined.

This function delegates to _sync_label_propagation for the actual label propagation process.

Currently, only synchronous label propagation is implemented. If algo.synchronous is false, the function will return nothing.

Example

julia> compute(FastLPA(), g)

GraphProperties.Communities.compute — Method

compute(algo::FastLPA, edge_list::Vector{Tuple{Int,Int}}, num_vertices::Int)

Execute the Fast Label Propagation algorithm using a precomputed edge list.

This variant of the compute function allows for direct input of a graph's edge list and number of vertices. It's particularly useful when the edge list has been precomputed or when working with a graph representation that doesn't conform to a SimpleGraph.

Arguments

algo::FastLPA: The Fast Label Propagation algorithm instance.

edge_list::Vector{Tuple{Int,Int}}: The edge list of the graph, where each edge is represented as a tuple of vertex indices.

num_vertices::Int: The number of vertices in the graph.

Returns

If algo.synchronous is true, it returns the result of synchronous label propagation; otherwise, it returns nothing.

Notes

The edge list should be lexicographically sorted and represent a valid graph.

Only synchronous label propagation is currently implemented.

GraphProperties.Communities.compute — Method

compute(algo::KClique, g::SimpleGraph)::Dict{Int, Int}

Detect communities in a graph g using the K-Clique algorithm.

The function first finds triangles (or 3-cliques) in the graph. It then constructs a k-clique graph where nodes represent triangles, and edges indicate overlap. The connected components of this k-clique graph give the communities in the original graph.

Arguments

algo::KClique: Indicates that the K-Clique algorithm should be used for community detection.

g::SimpleGraph: The graph on which to detect communities.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities

julia> g = generate(KarateClub())

julia> compute(KClique(), g)

Notes

Currently, the implementation is restricted to 3-cliques (triangles). Future versions might support other clique sizes.

GraphProperties.Communities.compute — Method

compute(algo::LabelPropagation, g::SimpleGraph)::LabelArray

Detect communities in a graph g using the Label Propagation algorithm.

The algorithm works by initially assigning each node a unique label. Then, in each iteration, each node adopts the label that is most frequent among its neighbors. The algorithm terminates when no node changes its label or after reaching a maximum number of iterations.

Arguments

algo::LabelPropagation: An instance indicating the settings of the Label Propagation algorithm.

g::SimpleGraph: The graph on which to detect communities.

Returns

A LabelArray where each index corresponds to a vertex and its value indicates its community label.

Example

julia> using GraphCommunities

julia> g = generate(KarateClub())

julia> communities = compute(LabelPropagation(), g)

Notes

The algorithm may not return the same community structure on different runs due to its heuristic nature. However, the structures should be reasonably similar and of comparable quality.

GraphProperties.Communities.compute — Method

compute(algo::Louvain, g::SimpleGraph)

Detect communities in a graph g using the Louvain algorithm, a method based on modularity optimization.

The algorithm consists of two phases that are repeated iteratively:

Local Phase: Each node is moved to the community that yields the highest modularity gain.

Aggregation Phase: A new graph is constructed where nodes represent communities from the previous phase.

These phases are repeated until the modularity ceases to increase significantly.

Arguments

algo::Louvain: Indicates that the Louvain algorithm should be used for community detection.

g::SimpleGraph: The graph on which to detect communities.

Returns

A dictionary mapping node IDs in the original graph to their respective community IDs.

Example

julia> using GraphCommunities

julia> g = generate(PlantedPartition())

julia> compute(Louvain(), g)

Notes

The algorithm may not return the same community structure on different runs due to its heuristic nature. However, the structures should be reasonably similar and of comparable quality.

GraphProperties.Communities.compute — Method

compute(algo::PageRank, g::AbstractGraph)::Vector{Float64}

Compute the PageRank values of the nodes in graph g using the specified PageRank algorithm.

Arguments

algo::PageRank: The PageRank algorithm configuration object. This should contain properties like (initial) damping factor (d), maximum number of iterations (max_iter), tolerance (tol), and the method: "classical", "iterative", or "adaptive" (method_name). The two latter methods are described in the paper PageRank: New Regularizations and Simulation Models (2011) by Boris T. Polyak and Anna V. Timonina

g::AbstractGraph: The graph for which to compute the PageRank. This can be a simple graph, directed or undirected, or a weighted version of these.

Returns

A vector of Float64 where each entry represents the PageRank value of the corresponding node in the graph.

Details

The function uses the power iteration method to find the principal eigenvector of an altered adjacency matrix. The altered matrix is created by 'damping' the original matrix, which is done by adding a constant to each entry as well as taking care of sinks. The iterative and adaptive methods increase the damping factor (decreases perturbance of P) as the algorithm progresses, which is shown to improve convergence. The iterative method increases the damping factor over a determined sequence, while the adaptive method maintains the damping factor until a relaxed tolerance is reached, then lowers the tolerance and repeats until the original tolerance is reached.

```julia julia> g = generate(PlantedPartition())

julia> algo = PageRank(d=0.85, maxiter=100, tol=1e-6, methodname="adaptive")

julia> compute(algo, g)

GraphProperties.Communities.draw_communities — Method

draw_communities(g::AbstractGraph, communities::Dict)

Draw the graph g with nodes colored based on their community assignments.

Arguments

g::AbstractGraph: The input graph.

communities::Dict: A dictionary mapping each vertex to its community.

Returns

A plot with nodes colored based on their community.

Note

This function will only work if each node in the graph is assigned to a community.

GraphProperties.Communities.draw_communities — Method

draw_communities(g::AbstractGraph, node_labels::Vector{Tuple{Int, Int}})

Draw the graph g with nodes colored based on their label assignments.

Arguments

g::AbstractGraph: The input graph.

node_labels::Vector{Tuple{Int, Int}}: A vector of tuples, each containing a node and its label.

Returns

A plot with nodes colored based on their labels.

Example

julia> g = generate(KarateClub())

julia> communities = compute(FastLPA(), g)

julia> draw_communities(g, communities)

Note

This function will only work if each node in the graph is included in the node_labels vector.

GraphProperties.Communities.find_triangles — Method

find_triangles(g::SimpleGraph)::Set{Set{Int}}

Find all the triangles in the given graph.

Arguments

g::SimpleGraph: The input graph to search for triangles.

Returns

A set containing all the triangles in the graph. Each triangle is represented as a set of 3 vertices.

GraphProperties.Communities.generate — Method

generate(structure::ChainedCliques)::SimpleGraph

Create a graph consisting of structure.r cliques, each of size structure.k, chained together.

structure.r represents the number of cliques.

structure.k represents the size of each clique.

Returns a SimpleGraph with the chained cliques.

GraphProperties.Communities.generate — Method

generate(structure::KarateClub)::SimpleGraph

Construct the famous Zachary's Karate Club graph. This graph represents the friendships between the 34 members of a karate club studied by Wayne W. Zachary in 1977.

Returns a SimpleGraph representing the Karate Club network.

GraphProperties.Communities.generate — Method

generate(structure::PlantedPartition)::SimpleGraph

Generate a graph based on the planted partition model.

structure.n_communities is the number of communities.

structure.nodes_per_community denotes the number of nodes per community.

structure.pintra is the probability of an edge within a community.

structure.pinter is the probability of an edge between communities.

Returns a SimpleGraph constructed based on the planted partition model.

GraphProperties.Communities.graph_modularity — Method

graph_modularity(g::SimpleGraph, node_to_community::Dict{Int, Int})::Number

Calculate the modularity of a graph g given a particular community assignment given by node_to_community.

Arguments

g::SimpleGraph: The input graph.

node_to_community::Dict{Int, Int}: A dictionary mapping each vertex to its community.

Returns

Float64: The modularity value.

GraphProperties.Communities.k_clique_graph — Method

k_clique_graph(triangles::Set{Set{Int}})::SimpleGraph

Construct a graph where each node represents a triangle from the input set, and edges are added between nodes if their respective triangles share two vertices.

Arguments

triangles::Set{Set{Int}}: A set of triangles, where each triangle is represented as a set of 3 vertices.

Returns

A graph with nodes representing triangles and edges based on shared vertices.

GraphProperties.Communities.label_propagation_sweep! — Method

label_propagation_sweep!(g::SimpleGraph, labels::LabelArray, algo::LabelPropagation,
                         X::VertexList, new_labels::LabelArray, label_counts::Vector{Int})

Perform a single iteration of the Label Propagation algorithm on the provided graph.

Arguments

g::SimpleGraph: The graph on which to detect communities.

labels::LabelArray: An array where each index corresponds to a vertex and its value to its community label.

algo::LabelPropagation: An instance indicating the settings of the Label Propagation algorithm.

X::VertexList: A list of vertices to be considered for the propagation in the given iteration.

new_labels::LabelArray: An array to store new labels if the algorithm is running in synchronous mode.

label_counts::LabelArray: An array to store counts of each label during propagation.

Notes

The function modifies the given labels array in-place, reflecting the label changes during propagation.

GraphProperties.Communities.labels_to_dict — Method

labels_to_dict(labels::Vector{Set{Int}})::Dict{Int, Int}

Convert a vector of sets to a dictionary mapping each element to its index in the vector.

GraphProperties.Communities.number_of_communities — Method

number_of_communities(communities::Vector{Int64})::Int

Calculate the number of unique communities in a graph.

This function takes a vector where each element represents the community assignment of a node in the graph and returns the total count of unique communities identified.

Arguments

communities::Vector{Int64}: A vector with each element representing the community ID assigned to a node in the graph.

Returns

Int: The number of unique communities found in the graph.

Example

julia> community_assignments = [1, 2, 1, 3, 2, 3, 3]

julia> number_of_communities(community_assignments)
3

GraphProperties.Communities.random_argmax — Method

random_argmax(arr::AbstractArray)::Int

Find all indices corresponding to the maximum value of the array and randomly return one of them.

Arguments

arr::AbstractArray: An array of values.

Returns

An index corresponding to a randomly selected maximum value of the array.

Notes

If multiple maxima exist in the array, one of them is chosen uniformly at random.

GraphProperties.Communities.update_label_counts! — Method

update_label_counts!(label_counts::Vector{Int}, labels::LabelArray, neighbors::VertexList)

Update the count of labels based on the neighbors of a given vertex.

Arguments

label_counts::LabelArray: An array to store the count of each label.

labels::LabelArray: An array where each index corresponds to a vertex and its value corresponds to its label.

neighbors::VertexList: List of neighboring vertices for the vertex being considered.

Notes

This function is used internally by label_propagation_sweep! to update the label counts for the vertices.