Degree Sequence Invariants

The DegreeSequenceInvariants module provides types and methods for analyzing invariants related to the degree sequence of a graph.

GraphProperties.DegreeSequenceInvariants.annihilation_number — Method

annihilation_number(g::SimpleGraph)

Return the annihilation number of the graph g.

The annihilation number of a graph g is the largest integer k such that the sum of the k smallest degrees in g is less than or equal to the number of edges in g.

Arguments

g::SimpleGraph: A simple graph.

Returns

The annihilation number of the graph g.

GraphProperties.DegreeSequenceInvariants.average_degree — Method

average_degree(g::AbstractGraph{T}) where T <: Integer

Return the (arithmetic) average degree of g.

GraphProperties.DegreeSequenceInvariants.havel_hakimi_residue — Method

havel_hakimi_residue(g::SimpleGraph)

Return the Havel-Hakimi residue of the graph g.

The Havel-Hakimi residue of a graph g is the length of the degree sequence after applying the Havel-Hakimi rule to the degree sequence of g. This method returns the length of the final sequence after applying the Havel-Hakimi rule as long as it may be applied.

Arguments

g::SimpleGraph: A simple graph.

Returns

The Havel-Hakimi residue of the graph g.

Example

julia> using Graphs

julia> using GraphProperties.DegreeSequenceInvariants

julia> g = PathGraph(5);

julia> havel_hakimi_residue(g)
2

GraphProperties.DegreeSequenceInvariants.maximum_degree — Method

maximum_degree(g::AbstractGraph{T}) where T <: Integer

Return the maximum degree of g.

GraphProperties.DegreeSequenceInvariants.minimum_degree — Method

minimum_degree(g::AbstractGraph{T},) where T <: Integer

Return the minimum degree of g.

GraphProperties.DegreeSequenceInvariants.slater_number — Method

slater_number(g::SimpleGraph{T},) where T <: Integer

Return the Slater invariant for the graph g.

The Slater invariant of a graph g is the sub-k-domination number of g where k = 1.

Arguments

g::SimpleGraph: A simple graph.

Returns

The Slater invariant of the graph g.

Example

julia> using Graphs

julia> using GraphProperties.DegreeSequenceInvariants

julia> g = PathGraph(5);

julia> slater_number(g)
2

GraphProperties.DegreeSequenceInvariants.sub_k_domination_number — Method

function sub_k_domination_number(
    g::SimpleGraph{T};
    k::Integer = 1,
) where T <: Integer

Return the sub-k-domination number of the graph g.

The sub-k-domination number of a graph g is the smallest integer t such that the sum of the degrees of the first t vertices in the sorted degree sequence with t is at least n. Scaling by 1/k. Please see the following reference for more information: D. Amos et al., "The sub-k-domination number of a graph with applications to k-domination", arXiv preprint arXiv:1611.02379, (2016)

Arguments

g::SimpleGraph: A simple graph.

k::Integer: The scaling factor.

Returns

The sub-k-domination number of the graph g.

Example

julia> using Graphs

julia> using GraphProperties.DegreeSequenceInvariants

julia> g = PathGraph(5);

julia> sub_k_domination_number(g, k=1)
2

GraphProperties.DegreeSequenceInvariants.sub_total_k_domination_number — Method

sub_total_k_domination_number(
    g::SimpleGraph{T};
    k::Integer = 1,
) where T <: Integer

Return the sub-total domination number of the graph g.

The sub-total domination number of a graph g is the smallest integer t such that the sum of the degrees of the first t vertices in the sorted degree sequence with t is at least k * n where n is the number of vertices in g and k is just used for scaling Please see the following reference for more information: R. Davila, "A note on sub-total domination in graphs", arXiv preprint arXiv:1701.07811, (2017)

Arguments

g::SimpleGraph: A simple graph.

k::Integer: The scaling factor.

Returns

The sub-total domination number of the graph g.

Example

julia> using Graphs

julia> using GraphProperties.DegreeSequenceInvariants

julia> g = CompleteGraph(6);

julia> sub_total_k_domination_number(g, k=1)
2