Draw

The Draw module provides resources for visualizing graphs and optimal sets.

GraphProperties.Draw.draw — Method

draw(g::AbstractGraph, optset::AbstractOptimalEdgeSet)

Draw a graph with the optimal edge set highlighted.

Arguments

g::AbstractGraph: The graph to draw.

optset::AbstractOptimalEdgeSet: The optimal edge set.

Returns

A plot of the graph with the optimal edge set highlighted.

GraphProperties.Draw.draw — Method

draw(g::AbstractGraph, optset::AbstractOptimalNodeSet)

Draw a graph with the optimal node set highlighted.

Arguments

g::AbstractGraph: The graph to draw.

optset::AbstractOptimalNodeSet: The optimal node set.

Returns

A plot of the graph with the optimal node set highlighted.