Graph IO

The GraphIO module includes functionality to load and write edgelist files

GraphProperties.GraphIO — Module

GraphIO

A module for reading and writing graphs from/to various file formats including .csv and .txt.

Provides functionality to:

Load a graph from an edge list file (load_edgelist)

Write a graph to an edge list file (write_edgelist)

GraphProperties.GraphIO.CSVFormat — Type

CSVFormat

A subtype of EdgeListFormat that represents the .csv file format.

GraphProperties.GraphIO.EdgeListFormat — Type

EdgeListFormat

An abstract type that represents the file format of an edge list.

Subtypes of EdgeListFormat include:

CSVFormat

TXTFormat

GraphProperties.GraphIO.TXTFormat — Type

TXTFormat

A subtype of EdgeListFormat that represents the .txt file format.

GraphProperties.GraphIO._determine_format — Method

_determine_format(path::String) -> EdgeListFormat

Determine the file format based on the file extension.

Arguments

path::String: The file path.

Returns

An instance of a subtype of EdgeListFormat indicating the file format.

GraphProperties.GraphIO._validate_node — Method

_va;idate_node(value, column::String)

Validate that a node is a positive integer.

GraphProperties.GraphIO.load_edgelist — Method

load_edgelist(path::String, format::CSVFormat)::SimpleGraph

Load a graph from a .csv edge list file.

Arguments

path::String: The path to the .csv file.

Returns

A SimpleGraph object constructed from the edge list.

GraphProperties.GraphIO.load_edgelist — Method

load_edgelist(path::String, format::TXTFormat)::SimpleGraph

Load a graph from a .txt edge list file.

Arguments

path::String: The path to the .txt file.

Returns

A SimpleGraph object constructed from the edge list.

GraphProperties.GraphIO.load_edgelist — Method

load_edgelist(path::String)::SimpleGraph

Load a graph from an edge list file (either .csv or .txt).

Arguments

path::String: The path to the edge list file.

Returns

A SimpleGraph object constructed from the edge list.

GraphProperties.GraphIO.write_edgelist! — Method

write_edgelist(g::SimpleGraph, path::String, format::CSVFormat)

Write a graph to a .csv edge list file.

Arguments

g::SimpleGraph: The graph object.

path::String: The path to save the .csv file.

GraphProperties.GraphIO.write_edgelist! — Method

write_edgelist!(g::SimpleGraph, path::String, format::TXTFormat)

Write a graph to a .txt edge list file.

Arguments

g::SimpleGraph: The graph object.

path::String: The path to save the .txt file.

GraphProperties.GraphIO.write_edgelist! — Method

write_edgelist(g::SimpleGraph, path::String)

Write a graph to an edge list file (either .csv or .txt).

Arguments

g::SimpleGraph: The graph object.

path::String: The path to save the edge list file.