Graph Rules

The GraphRules module contains types and methods built for applying rules in-place.

GraphProperties.GraphRules.AbstractDegreeSequenceRule — Type

AbstractDegreeSequenceRule

An abstract type that represents a rule that is applied to a degree sequence.

GraphProperties.GraphRules.AbstractGraphRule — Type

AbstractGraphRule

An abstract type that represents a rule that is applied to a graph.

GraphProperties.GraphRules.DominationRule — Type

DominationRule

A rule that is applied to a graph to determine if vertices set S is a dominating set.

GraphProperties.GraphRules.HavelHakimiRule — Type

HavelHakimiRule

A rule that is applied to a degree sequence iteratively to determine if it is graphical.

GraphProperties.GraphRules.ZeroForcingRule — Type

ZeroForcingRule

A rule to determine if vertices set blue is a zero forcing set of g.

GraphProperties.GraphRules.apply! — Method

apply!(::Type{DominationRule}, target_set::Set{T}, g::SimpleGraph{T}) where T <: Integer

Apply the domination rule to the target set of vertices in the graph g.

The domination rule states that if a vertex v is in the target set, then on the next iteration (after applying the rule) all of the neighbors of v will be in the target set. The method modifies the target set in place.

Arguments

target_set::Set{T}: The set of vertices to which the rule is applied.

g::SimpleGraph{T}: The graph to which the rule is applied.

Example

julia> using Graphs

julia> using GraphProperties.GraphRules

julia> g = path_graph(5)
{5, 4} undirected simple Int64 graph

julia> target_set = Set([1]);

julia> apply!(DominationRule, target_set, g)

julia> target_set
Set([1, 2])

julia> apply!(DominationRule, target_set, g)

julia> target_set
Set([1, 2, 3])

GraphProperties.GraphRules.apply! — Method

apply!(::Type{HavelHakimiRule}, sequence::Vector{Int}, eliminations::Vector{Int})

Apply the Havel-Hakimi rule to the given sequence.

The Havel-Hakimi rule is a graphical sequence test that determines if a given sequence is graphical. One step of the rule is to remove the largest number, d1, from the sequence and subtract 1 from the next d1 numbers in the sorted sequence. If the sequence is all zeros, then the original sequence is graphical. If the rule cannot be applied to the sequence, then the sequence is not graphical. This method applies one iteration of the Havel-Hakimi rule to a given sequence in-place and returns true if the rule was applied successfully. Otherwise, a DomainError is thrown. Note: if the sequence is not already sorted in non-increasing order, it will be sorted in-place.

Arguments

sequence::Vector{Int}: A sequence of integers.

eliminations::Vector{Int}: A vector to store the eliminations.

Returns

true if the rule was applied successfully, otherwise a DomainError is thrown.

Example

julia> using GraphInvariants

julia> D = [3, 3, 2, 2, 1, 1];

julia> eliminations = Int[];

julia> apply!(HavelHakimiRule, D, eliminations)
true

julia> D
5-element Vector{Int64}:
 2
 1
 1
 1
 1

julia> eliminations
1-element Vector{Int64}:
 3

GraphProperties.GraphRules.apply! — Method

apply!(::Type{ZeroForcingRule}, blue::Set{T}, g::SimpleGraph{T}; max_iter=100_000)

Iteratively applies the zero forcing rule to g with the set of blue vertices blue.

The zero forcing rule is a rule that is applied to a graph g with a set of blue vertices blue. The rule is applied iteratively to the graph by checking each vertex in the graph. If a vertex is blue and has exactly one white neighbor, then the white neighbor is turned blue. The rule is applied until no more changes can be made or until max_iter iterations have been made.

Arguments

blue::Set{T}: A set of vertices that are blue.

g::SimpleGraph{T}: A simple graph.

max_iter::Int: The maximum number of iterations to apply the rule.

Example

julia> using Graphs

julia> using GraphProperties.GraphRules

julia> g = PathGraph(5);

julia> blue = Set([1, 4]);

julia> apply!(ZeroForcingRule, blue, g)

julia> blue
Set{Int64} with 5 elements:
  5
  4
  2
  3
  1