Invariants

The Invariants module provides functions to compute invariants for a given graph. These invariants can categorized into these general categories:

Chromatic Colorings

Cliques

Domination

Independence

Matching

Zero Forcing

Methods use several optimization approaches including integer programming which is handled through JuMP and uses the default optimizer HiGHS.

GraphProperties.Invariants — Module

module Invariants

The Invariants module provides functionalities for computing various graph invariants and optimal sets, such as cliques, colorings, dominations, and matchings. It includes advanced features for degree sequence invariants and graphical operations based on the rules defined within the module.

Exported Types

AbstractOptimalNodeSet: Represents sets of nodes that meet an optimality criterion.

AbstractOptimalEdgeSet: Represents sets of edges that meet an optimality criterion.

AbstractOptimalNodeColoring: Represents an optimal coloring of nodes in a graph.

AbstractOptimalEdgeColoring: Represents an optimal coloring of edges in a graph.

AbstractCardinality: Represents the cardinality of a set.

AbstractOptimalCardinality: Represents the cardinality of an optimal set.

AbstractDegreeSequenceInvariant: Represents an invariant of a graph's degree sequence.

Exported Functions

compute: Computes the optimal set or invariant of a graph.

Usage

The module allows for the computation of complex graph invariants and operations through high-level functions and types. The types are used to represent the results of these computations, while the functions are used to perform them.

Examples

julia> using GraphProperties.Invariants: compute, MaximumIndependentSet

julia> α = compute(MaximumIndependentSet, my_graph)

GraphProperties.Invariants.AbstractCardinality — Type

AbstractCardinality

Represents the cardinality of a set.

GraphProperties.Invariants.AbstractDegreeSequenceInvariant — Type

AbstractDegreeSequenceInvariant

Represents an invariant of a graph's degree sequence.

GraphProperties.Invariants.AbstractOptimalCardinality — Type

AbstractOptimalCardinality

Represents the cardinality of an optimal set.

GraphProperties.Invariants.AbstractOptimalEdgeColoring — Type

AbstractOptimalEdgeColoring

Represents an optimal coloring of edges in a graph.

GraphProperties.Invariants.AbstractOptimalEdgeSet — Type

AbstractOptimalEdgeSet

Represents a set of edges that meet an optimality criterion.

GraphProperties.Invariants.AbstractOptimalNodeColoring — Type

AbstractOptimalNodeColoring

Represents an optimal coloring of nodes in a graph.

GraphProperties.Invariants.AbstractOptimalNodeSet — Type

AbstractOptimalNodeSet

Represents a set of nodes that meet an optimality criterion.

GraphProperties.Invariants.AnnihilationNumber — Type

AnnihilationNumber

Represents the annihilation number of a graph.

GraphProperties.Invariants.AverageDegree — Type

AverageDegree

Represents the average degree of a graph.

GraphProperties.Invariants.ChromaticNumber — Type

ChromaticNumber

Represents the number of colors in a minimum proper coloring of a graph.

GraphProperties.Invariants.CliqueNumber — Type

CliqueNumber

Represents the cardinality of a maximum clique of a graph.

GraphProperties.Invariants.Diameter — Type

Diameter

Represents the diameter of a graph.

GraphProperties.Invariants.DominationNumber — Type

DominationNumber

Represents the cardinality of a minimum dominating set of a graph.

GraphProperties.Invariants.EdgeDominationNumber — Type

EdgeDominationNumber

Represents the cardinality of a minimum edge dominating set of a graph.

GraphProperties.Invariants.Girth — Type

Girth

Represents the girth of a graph.

GraphProperties.Invariants.HavelHakimiResidue — Type

HavelHakimiResidue

Represents the Havel-Hakimi residue of a graph.

GraphProperties.Invariants.IndependenceNumber — Type

IndependenceNumber

Represents the cardinality of a maximum independent set of a graph.

GraphProperties.Invariants.IndependentDominationNumber — Type

IndependentDominationNumber

Represents the cardinality of a minimum independent dominating set of a graph.

GraphProperties.Invariants.LocatingDominationNumber — Type

LocatingDominationNumber

Represents the cardinality of a minimum locating dominating set of a graph.

GraphProperties.Invariants.MatchingNumber — Type

MatchingNumber

Represents the cardinality of a maximum matching of a graph.

GraphProperties.Invariants.MaximumClique — Type

MaximumClique

Represents a maximum clique of a graph.

GraphProperties.Invariants.MaximumDegree — Type

MaximumDegree

Represents the maximum degree of a graph.

GraphProperties.Invariants.MaximumIndependentSet — Type

MaximumIndependentSet

Represents a maximum independent set of a graph.

GraphProperties.Invariants.MaximumMatching — Type

MaximumMatching

Represents a maximum matching of a graph.

GraphProperties.Invariants.MinimumDegree — Type

MinimumDegree

Represents the minimum degree of a graph.

GraphProperties.Invariants.MinimumDominatingSet — Type

MinimumDominatingSet

Represents a minimum dominating set of a graph.

GraphProperties.Invariants.MinimumEdgeDominatingSet — Type

MinimumEdgeDominatingSet

Represents a minimum edge dominating set of a graph.

GraphProperties.Invariants.MinimumIndependentDominatingSet — Type

MinimumIndependentDominatingSet

Represents a minimum independent dominating set of a graph.

GraphProperties.Invariants.MinimumLocatingDominatingSet — Type

MinimumLocatingDominatingSet

Represents a minimum locating dominating set of a graph.

GraphProperties.Invariants.MinimumPairedDominatingSet — Type

MinimumPairedDominatingSet

Represents a minimum paired dominating set of a graph.

GraphProperties.Invariants.MinimumPowerDominatingSet — Type

MinimumPowerDominatingSet

Represents a minimum power dominating set of a graph.

GraphProperties.Invariants.MinimumProperColoring — Type

MinimumProperColoring

Represents a minimum proper coloring of a graph.

GraphProperties.Invariants.MinimumTotalDominatingSet — Type

MinimumTotalDominatingSet

Represents a minimum total dominating set of a graph.

GraphProperties.Invariants.MinimumZeroForcingSet — Type

MinimumZeroForcingSet

Represents a minimum zero forcing set of a graph.

GraphProperties.Invariants.Order — Type

Order

Represents the number of nodes in a graph.

GraphProperties.Invariants.PairedDominationNumber — Type

PairedDominationNumber

Represents the cardinality of a minimum paired dominating set of a graph.

GraphProperties.Invariants.PowerDominationNumber — Type

PowerDominationNumber

Represents the cardinality of a minimum power dominating set of a graph.

GraphProperties.Invariants.Radius — Type

Radius

Represents the radius of a graph.

GraphProperties.Invariants.Size — Type

Size

Represents the number of edges in a graph.

GraphProperties.Invariants.SlaterNumber — Type

SlaterNumber

Represents the Slater number of a graph.

GraphProperties.Invariants.SubKDominationNumber — Type

SubKDominationNumber

Represents the sub-k domination number of a graph.

GraphProperties.Invariants.SubTotalDominationNumber — Type

SubTotalDominationNumber

Represents the sub-total domination number of a graph.

GraphProperties.Invariants.TotalDominationNumber — Type

TotalDominationNumber

Represents the cardinality of a minimum total dominating set of a graph.

GraphProperties.Invariants.ZeroForcingNumber — Type

ZeroForcingNumber

Represents the cardinality of a minimum zero forcing set of a graph.

Base.:== — Method

==(
    a::AbstractOptimalEdgeSet,
    b::AbstractOptimalEdgeSet
)::Bool

Determine if two sets of edges are equivalent in an undirected graph context.

This method treats each edge as undirected, meaning the order of vertices in each edge tuple or Graphs.SimpleGraphs.SimpleEdge is ignored. Two sets of edges are considered equal if they contain the same pairs of vertices, regardless of the order in which the vertices appear in each edge.

Implementation Notes

It does not matter if the input sets of edges are of different concrete types, as long as they are both subtypes of AbstractOptimalEdgeSet. See the example below.

Arguments

a::AbstractOptimalEdgeSet: The first set of edges to compare.

b::AbstractOptimalEdgeSet: The second set of edges to compare.

Returns

true if both sets of edges represent the same undirected connections, false otherwise.

Examples

julia> using Graphs.SimpleGraphs: SimpleEdge

julia> using GraphProperties.Invariants

julia> a = MaximumMatching([SimpleEdge(1, 2), SimpleEdge(2, 3)]);

julia> b = MinimumEdgeDominatingSet([SimpleEdge(2, 1), SimpleEdge(3, 2)]);

julia> a == b
true

GraphProperties.Invariants.clique_number — Method

function clique_number(g::AbstractGraph; optimizer=HiGHS.Optimizer)

Return the clique number of g.

A clique of a graph g is a subset of the vertices of g such that every two distinct vertices in the clique are adjacent. The clique number of g is the cardinality of a maximum clique of g.

Implementation Notes

Computes the independence number of the complement of g.

Arguments

g::AbstractGraph: The graph to compute the clique number of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the maximum independent set problem.

Returns

The clique number of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = path_graph(5)
{5, 4} undirected simple Int64 graph

julia> clique_number(g)
2

GraphProperties.Invariants.compute — Method

compute(::Type{BruteForceMinimumZeroForcingSet}, g::AbstractGraph)

Return a minimum zero forcing set of g using a brute force method.

This method assures correctness, but is not efficient for large graphs. A set of nodes is considered zero forcing if it can force all nodes in the graph to be colored by the following rules: - If a node is colored, it remains colored. - If a node is colored and has exactly one uncolored neighbor, that neighbor is colored. A minimum zero forcing set is a zero forcing set of minimum cardinality.

Arguments

g::AbstractGraph: The input graph.

Returns

A minimum zero forcing set of g.

Examples

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = path_graph(5)
{5, 4} undirected simple Int64 graph

julia> compute(BruteForceMinimumZeroForcingSet, g)
BruteForceMinimumZeroForcingSet([1])

julia> h = Graphs.complete_graph(4)
{4, 6} undirected simple Int64 graph

julia> compute(BruteForceMinimumZeroForcingSet, h)
BruteForceMinimumZeroForcingSet([1, 2, 3])

GraphProperties.Invariants.compute — Method

compute(::Type{MaximumClique}, g::AbstractGraph{T}; optimizer=HiGHS.Optimizer)

Return the maximum clique of g.

A clique of a graph g is a subset of the vertices of g such that every two distinct vertices in the clique are adjacent. A maximum clique is a clique of the largest possible size.

Arguments

g::AbstractGraph{T}: The graph to compute the maximum clique of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the maximum clique problem.

Returns

A MaximumClique object representing the maximum clique of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> compute(MaximumClique, g)
MaximumClique([1, 5])

GraphProperties.Invariants.compute — Method

function compute(
    ::Type{MaximumIndependentSet},
    g::AbstractGraph{T};
    optimizer=Cbc.Optimizer,
) where T <: Integer

Obtain a maximum independent set of g.

Implementation Notes

The independent set is found by solving the following linear program:

\[\begin{align*} \max_{i \in \mathbb{Z}} & \sum\limits_{v \in V} x_v \\ \;\;\text{s.t.}\; & x_v \in \{0, 1\} & \forall v \in V \\ & x_u + x_v \leq 1 & \forall uv \in E \end{align*}\]

Arguments

g::AbstractGraph{T}: The graph to compute the maximum independent set of.

Keywords

optimizer=Cbc.Optimizer: The optimizer to use to solve the maximum independent set problem.

Returns

A MaximumIndependentSet object representing the maximum independent set of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = path_graph(5)
{5, 4} undirected simple Int64 graph

julia> compute(MaximumIndependentSet, g)
MaximumIndependentSet([1, 3, 5])

GraphProperties.Invariants.compute — Method

compute(::Type{MaximumMatching}, g::SimpleGraph{T}; optimizer=HiGHS.Optimizer) where T <: Integer

Return a maximum matching of g.

A matching of a graph g is a subset of the edges of g such that no two edges share a common vertex. A matching is said to be maximum if it has the largest possible cardinality among all matchings of g.

Arguments

g::SimpleGraph{T}: The graph to compute a maximum matching of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the maximum matching problem.

Returns

A maximum matching of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> compute(MaximumMatching, g)
MaximumMatching(Graphs.SimpleGraphs.SimpleEdge{Int64}[Edge 1 => 2, Edge 3 => 4])

GraphProperties.Invariants.compute — Method

compute(
    ::Type{MinimumDominatingSet},
    g::AbstractGraph{T};
    optimizer=HiGHS.Optimizer
) where T <: Integer

Return a minimum dominating set of g.

A dominating set of a graph g is a subset S of the vertices of g such that every vertex not in S is adjacent to at least one member of S. A dominating set is said to be minimum if it has the smallest possible cardinality among all dominating sets of g.

Arguments

g::AbstractGraph{T}: The graph to compute the minimum dominating set of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the minimum dominating set problem.

Returns

A MinimumDominatingSet object representing the minimum dominating set of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> compute(MinimumDominatingSet, g)
MinimumDominatingSet([2, 5])

GraphProperties.Invariants.compute — Method

compute(::Type{MinimumEdgeDominatingSet}, g::SimpleGraph{T}; optimizer=HiGHS.Optimizer)

Return a minimum edge dominating set of g.

An edge dominating set of a graph g is a subset S of the edges of g such that every edge in g is either in S or adjacent to an edge in S. A minimum edge dominating set is an edge dominating set of smallest possible cardinality.

Arguments

g::SimpleGraph{T}: The graph to compute the minimum edge dominating set of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the optimization problem.

Returns

A MinimumEdgeDominatingSet object representing the minimum edge dominating set of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> compute(MinimumEdgeDominatingSet, g)
MinimumEdgeDominatingSet(Graphs.SimpleGraphs.SimpleEdge{Int64}[Edge 1 => 5, Edge 2 => 3])

GraphProperties.Invariants.compute — Method

compute(::Type{MinimumZeroForcingSet}, g::SimpleGraph)

Return a minimum zero forcing set of g.

GraphProperties.Invariants.compute — Method

compute(
    ::Type{MinimumProperColoring},
    g::SimpleGraph{T};
    optimizer = HiGHS.Optimizer
) where T <: Integer

Return the minimum proper coloring of g by solving a Binary Program.

Arguments

g::SimpleGraph{T}: The graph to compute the minimum proper coloring of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the minimum proper coloring problem.

Returns

A MinimumProperColoring object representing the minimum proper coloring of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = SimpleGraph(5)
{5, 0} undirected simple Int64 graph

julia> add_edge!(g, 1, 2);

julia> add_edge!(g, 1, 3);

julia> add_edge!(g, 1, 4);

julia> add_edge!(g, 1, 5);

julia> compute(MinimumProperColoring, g)
MinimumProperColoring(Dict(5 => 2, 4 => 2, 2 => 2, 3 => 2, 1 => 1))

GraphProperties.Invariants.compute — Method

compute(
    ::Type{MinimumTotalDominatingSet},
    g::SimpleGraph{T};
    optimizer=HiGHS.Optimizer
) where T <: Integer

Return a minimum total dominating set of g.

A total dominating set of a graph g is a subset S of the vertices of g such that every vertex in g is adjacent to at least one member of S. A total dominating set is said to be minimum if it has the smallest possible cardinality among all total dominating sets of g.

Arguments

g::SimpleGraph{T}: The graph to compute the minimum total dominating set of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the optimization problem.

Returns

A MinimumTotalDominatingSet object representing the minimum total dominating set of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> compute(MinimumTotalDominatingSet, g)
MinimumTotalDominatingSet([2, 3, 4])

GraphProperties.Invariants.compute — Method

compute(::Type{MinimumZeroForcingSet}, g::SimpleGraph)

Return a minimum zero forcing set of the graph g

This method uses an Integer Programming formulation from Brimkov et al. (2019), published in EJOR.

Implementation Notes

This method uses the HiGHS optimizer to solve the Integer Program by deafault and has encountered some issues with this optimizer. If you encounter issues, try using a different optimizer.

Arguments

g::DiGraph{T}: The input graph, with directed edges.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the minimum zero forcing set problem.

Example

julia> using Graphs

julia> h = Graphs.complete_graph(4)
{4, 6} undirected simple Int64 graph

julia> directed_h = Graphs.DiGraph(h)
{4, 12} directed simple Int64 graph

julia> compute(MinimumZeroForcingSet, directed_h)
MinimumZeroForcingSet(Set{Int64}([1, 2, 3]))

GraphProperties.Invariants.compute — Method

compute(::Type{MinimumZeroForcingSet}, g::SimpleGraph)

Return a minimum zero forcing set of the graph g

This method uses an Integer Programming formulation from Brimkov et al. (2019), published in EJOR.

Implementation Notes

This method converts the input graph to a directed graph, replacing each edge with two directed edges. If the input graph is already directed, a different method is used. This method uses the HiGHS optimizer to solve the Integer Program by deafault and has encountered some issues with this optimizer. If you encounter issues, try using a different optimizer.

Arguments

g::SimpleGraph: The input graph.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the minimum zero forcing set problem.

Example

julia> using Graphs

julia> h = Graphs.complete_graph(4)
{4, 6} undirected simple Int64 graph

julia> compute(MinimumZeroForcingSet, h)
MinimumZeroForcingSet(Set{Int64}([1, 2, 3]))

GraphProperties.Invariants.domination_number — Method

domination_number(g::SimpleGraph; optimizer=HiGHS.Optimizer)})

Return the domination number of g.

A dominating set of a graph g is a subset S of the vertices of g such that every vertex in g is either in S or adjacent to a vertex in S. The domination number of g is the cardinality of a minimum dominating set of g.

Arguments

g::SimpleGraph{T}: The graph to compute the domination number of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the minimum dominating set problem.

Returns

The domination number of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> domination_number(g)
2

GraphProperties.Invariants.edge_domination_number — Method

edge_domination_number(g::SimpleGraph{T}; optimizer=HiGHS.Optimizer)

Return the edge domination number of g.

An edge dominating set of a graph g is a subset S of the edges of g such that every edge in g is incident to at least one edge in S. The edge domination number of g is the cardinality of a minimum edge dominating set of g.

Arguments

g::SimpleGraph{T}: The graph to compute the edge domination number of.

Returns

The edge domination number of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> edge_domination_number(g)
2

GraphProperties.Invariants.independence_number — Method

function independence_number(
    g::SimpleGraph{T};
    optimizer=Cbc.Optimizer,
) where T <: Integer

Return the independence number of g.

Implementation Notes

This function relies on compute to obtain a maximum independent set of g. The optimizer is passed to max_independent_set.

Arguments

g::SimpleGraph{T}: The graph to compute the independence number of.

Keywords

optimizer=Cbc.Optimizer: The optimizer to use to solve the maximum independent set problem.

Returns

The independence number of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> independence_number(g)
2

GraphProperties.Invariants.independent_domination_number — Method

compute(::Type{IndependentDominationNumber}, g::SimpleGraph{T}; optimizer=HiGHS.Optimizer)

Return the independent domination number of g.

An independent dominating set of a graph g is a subset S of the vertices of g such that every vertex in g is either in S or adjacent to a vertex in S, and S forms an independent set in g. The independent domination number of g is the cardinality of a minimum independent dominating set of g.

Arguments

g::SimpleGraph{T}: The graph to compute the independent domination number of.

Returns

The independent domination number of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> independent_domination_number(g)
2

GraphProperties.Invariants.matching_number — Method

matching_number(g::SimpleGraph; optimizer=HiGHS.Optimizer)

Return the matching number of g.

A matching of a graph g is a subset of the edges of g such that no two edges in the matching share a vertex. The matching number of g is the cardinality of a maximum matching of g.

Arguments

g::SimpleGraph: The graph to compute the matching number of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the maximum matching problem.

Returns

The matching number of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> matching_number(g)
2

GraphProperties.Invariants.power_domination_number — Method

power_domination_number(g::SimpleGraph{T})

Return the power domination number of g.

GraphProperties.Invariants.total_domination_number — Method

total_domination_number(g::SimpleGraph{T}; optimizer=HiGHS.Optimizer)

Return the total domination number of g.

A total dominating set of a graph g is a subset S of the vertices of g such that every vertex in g is adjacent to at least one vertex in S. The total domination number of g is the cardinality of a minimum total dominating set of g.

Arguments

g::SimpleGraph{T}: The graph to compute the total domination number of.

Keywords

optimizer=HiGHS.Optimizer: The optimizer to use to solve the minimum total dominating set problem.

Returns

The total domination number of g.

Example

julia> using Graphs

julia> using GraphProperties.Invariants

julia> g = cycle_graph(5)
{5, 5} undirected simple Int64 graph

julia> total_domination_number(g)
3

GraphProperties.Invariants.zero_forcing_number — Method

zero_forcing_number(g::AbstractGraph{T}) where T <: Integer

Return the zero forcing number of g.

A set of nodes in a graph g is a zero forcing set if it can force all nodes in the graph to be colored by the following rules: - If a node is colored, it remains colored. - If a node is colored and has exactly one uncolored neighbor, that neighbor is colored. A minimum zero forcing set is a zero forcing set of minimum cardinality. The zero forcing number of g is the cardinality of a minimum zero forcing set of g.

###Implementation Notes This method uses an Integer Programming formulation from Brimkov et al. (2019). For a brute force method, see compute(BruteForceMinimumZeroForcingSet, g).

Arguments

g::AbstractGraph{T}: The graph to compute the zero forcing number of.

Returns

The zero forcing number of g.