GraphProperties

Documentation for GraphProperties.

GraphProperties.Arboricity

GraphProperties.BipartiteDimension

GraphProperties.Boxicity

GraphProperties.Choosability

GraphProperties.ChromaticIndex

GraphProperties.ChromaticNumber

GraphProperties.CliqueNumber

GraphProperties.CliqueWidth

GraphProperties.Degeneracy

GraphProperties.DegreeSequence

GraphProperties.DigraphIsDAG

GraphProperties.DigraphIsOrientation

GraphProperties.DigraphIsPlanar

GraphProperties.DigraphIsStronglyConnected

GraphProperties.DigraphIsWeaklyConnected

GraphProperties.DominationNumber

GraphProperties.EdgeConnectivity

GraphProperties.EdgeCoverNumber

GraphProperties.FeedbackVertexSetNumber

GraphProperties.FractionalChromaticNumber

GraphProperties.FractionalMatchingNumber

GraphProperties.Girth

GraphProperties.GraphIsBipartite

GraphProperties.GraphIsChordal

GraphProperties.GraphIsCircleGraph

GraphProperties.GraphIsCograph

GraphProperties.GraphIsComparabilityGraph

GraphProperties.GraphIsComplete

GraphProperties.GraphIsConnected

GraphProperties.GraphIsCycle

GraphProperties.GraphIsDistanceHereditary

GraphProperties.GraphIsForest

GraphProperties.GraphIsIndifferenceGraph

GraphProperties.GraphIsIntervalGraph

GraphProperties.GraphIsMeynielGraph

GraphProperties.GraphIsPath

GraphProperties.GraphIsPerfect

GraphProperties.GraphIsPermutationGraph

GraphProperties.GraphIsPlanar

GraphProperties.GraphIsPtolemaic

GraphProperties.GraphIsRegular

GraphProperties.GraphIsSplitGraph

GraphProperties.GraphIsTree

GraphProperties.GraphIsTriangleFree

GraphProperties.GraphIsTriviallyPerfect

GraphProperties.GraphProperty

GraphProperties.HadwigerNumber

GraphProperties.IndependenceNumber

GraphProperties.IntersectionNumber

GraphProperties.IsDirectedGraph

GraphProperties.IsUndirectedGraph

GraphProperties.MatchingNumber

GraphProperties.MaximumDegree

GraphProperties.MinimumDegree

GraphProperties.NumberOfArcs

GraphProperties.NumberOfConnectedComponents

GraphProperties.NumberOfEdges

GraphProperties.NumberOfVertices

GraphProperties.Pathwidth

GraphProperties.PropertyComparison

GraphProperties.Sphericity

GraphProperties.Splittance

GraphProperties.Strength

GraphProperties.StrongDominationNumber

GraphProperties.Treewidth

GraphProperties.TwinWidth

GraphProperties.VertexConnectivity

GraphProperties.VertexCoverNumber

GraphProperties.GraphProperty — Type

GraphProperty::Type

Abstract supertype for graph properties.

The only type parameter is the type of each value of the property.

GraphProperties.PropertyComparison — Type

PropertyComparison(comparison, property, value)

Turn a GraphProperty{<:Real} into a GraphProperty{Bool} by attaching a comparison against a value.

GraphProperties.Arboricity — Type

Arboricity()

The arboricity of an undirected graph.

julia> using GraphProperties

julia> Arboricity() isa GraphProperty{Integer}
true

GraphProperties.BipartiteDimension — Type

BipartiteDimension()

The bipartite dimension of an undirected graph. Also known as the biclique cover number.

julia> using GraphProperties

julia> BipartiteDimension() isa GraphProperty{Integer}
true

GraphProperties.Boxicity — Type

Boxicity()

The boxicity of an undirected graph.

julia> using GraphProperties

julia> Boxicity() isa GraphProperty{Integer}
true

GraphProperties.Choosability — Type

Choosability()

The choosability of an undirected graph. Also known as the list colorability or as the list chromatic number.

julia> using GraphProperties

julia> Choosability() isa GraphProperty{Integer}
true

GraphProperties.ChromaticIndex — Type

ChromaticIndex()

The chromatic index of an undirected graph. Also known as the edge chromatic number.

julia> using GraphProperties

julia> ChromaticIndex() isa GraphProperty{Integer}
true

GraphProperties.ChromaticNumber — Type

ChromaticNumber()

The chromatic number of an undirected graph.

julia> using GraphProperties

julia> ChromaticNumber() isa GraphProperty{Integer}
true

GraphProperties.CliqueNumber — Type

CliqueNumber()

The clique number of an undirected graph.

julia> using GraphProperties

julia> CliqueNumber() isa GraphProperty{Integer}
true

GraphProperties.CliqueWidth — Type

CliqueWidth()

The clique-width of an undirected graph.

julia> using GraphProperties

julia> CliqueWidth() isa GraphProperty{Integer}
true

GraphProperties.Degeneracy — Type

Degeneracy()

The degeneracy of an undirected graph. Also known as the width or as the linkage.

julia> using GraphProperties

julia> Degeneracy() isa GraphProperty{Integer}
true

GraphProperties.DegreeSequence — Type

DegreeSequence()

The degree sequence of an undirected graph.

julia> using GraphProperties

julia> DegreeSequence() isa GraphProperty{AbstractVector}
true

GraphProperties.DigraphIsDAG — Type

DigraphIsDAG()

Is a directed graph acyclic (a DAG)?

julia> using GraphProperties

julia> DigraphIsDAG() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsOrientation — Type

DigraphIsOrientation()

Is a directed graph an orientation?

julia> using GraphProperties

julia> DigraphIsOrientation() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsPlanar — Type

DigraphIsPlanar()

Is a directed graph planar?

julia> using GraphProperties

julia> DigraphIsPlanar() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsStronglyConnected — Type

DigraphIsStronglyConnected()

Is a directed graph strongly connected?

julia> using GraphProperties

julia> DigraphIsStronglyConnected() isa GraphProperty{Bool}
true

GraphProperties.DigraphIsWeaklyConnected — Type

DigraphIsWeaklyConnected()

Is a directed graph weakly connected?

julia> using GraphProperties

julia> DigraphIsWeaklyConnected() isa GraphProperty{Bool}
true

GraphProperties.DominationNumber — Type

DominationNumber()

The domination number of an undirected graph.

julia> using GraphProperties

julia> DominationNumber() isa GraphProperty{Integer}
true

GraphProperties.EdgeConnectivity — Type

EdgeConnectivity()

The edge-connectivity of an undirected graph.

julia> using GraphProperties

julia> EdgeConnectivity() isa GraphProperty{Integer}
true

GraphProperties.EdgeCoverNumber — Type

EdgeCoverNumber()

The edge cover number of an undirected graph.

julia> using GraphProperties

julia> EdgeCoverNumber() isa GraphProperty{Integer}
true

GraphProperties.FeedbackVertexSetNumber — Type

FeedbackVertexSetNumber()

The feedback vertex set number of a graph.

julia> using GraphProperties

julia> FeedbackVertexSetNumber() isa GraphProperty{Integer}
true

GraphProperties.FractionalChromaticNumber — Type

FractionalChromaticNumber()

The fractional chromatic number of an undirected graph.

julia> using GraphProperties

julia> FractionalChromaticNumber() isa GraphProperty{Real}
true

GraphProperties.FractionalMatchingNumber — Type

FractionalMatchingNumber()

The fractional matching number of an undirected graph.

julia> using GraphProperties

julia> FractionalMatchingNumber() isa GraphProperty{Real}
true

GraphProperties.Girth — Type

Girth()

The girth of a undirected graph: the length of its shortest cycle.

julia> using GraphProperties

julia> Girth() isa GraphProperty{Integer}
true

GraphProperties.GraphIsBipartite — Type

GraphIsBipartite()

Is an undirected graph bipartite?

julia> using GraphProperties

julia> GraphIsBipartite() isa GraphProperty{Bool}
true

GraphProperties.GraphIsChordal — Type

GraphIsChordal()

Is an undirected graph chordal?

julia> using GraphProperties

julia> GraphIsChordal() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCircleGraph — Type

GraphIsCircleGraph()

Is an undirected graph a circle graph?

julia> using GraphProperties

julia> GraphIsCircleGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCograph — Type

GraphIsCograph()

Is an undirected graph a cograph?

julia> using GraphProperties

julia> GraphIsCograph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsComparabilityGraph — Type

GraphIsComparabilityGraph()

Is an undirected graph a comparability graph?

julia> using GraphProperties

julia> GraphIsComparabilityGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsComplete — Type

GraphIsComplete()

Is an undirected graph complete?

julia> using GraphProperties

julia> GraphIsComplete() isa GraphProperty{Bool}
true

GraphProperties.GraphIsConnected — Type

GraphIsConnected()

Is an undirected graph connected?

julia> using GraphProperties

julia> GraphIsConnected() isa GraphProperty{Bool}
true

GraphProperties.GraphIsCycle — Type

GraphIsCycle()

Is an undirected graph a cycle?

julia> using GraphProperties

julia> GraphIsCycle() isa GraphProperty{Bool}
true

GraphProperties.GraphIsDistanceHereditary — Type

GraphIsDistanceHereditary()

Is an undirected graph distance-hereditary?

julia> using GraphProperties

julia> GraphIsDistanceHereditary() isa GraphProperty{Bool}
true

GraphProperties.GraphIsForest — Type

GraphIsForest()

Is an undirected graph a forest?

julia> using GraphProperties

julia> GraphIsForest() isa GraphProperty{Bool}
true

GraphProperties.GraphIsIndifferenceGraph — Type

GraphIsIndifferenceGraph()

Is an undirected graph an indifference graph?

julia> using GraphProperties

julia> GraphIsIndifferenceGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsIntervalGraph — Type

GraphIsIntervalGraph()

Is an undirected graph an interval graph?

julia> using GraphProperties

julia> GraphIsIntervalGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsMeynielGraph — Type

GraphIsMeynielGraph()

Is an undirected graph a Meyniel graph?

julia> using GraphProperties

julia> GraphIsMeynielGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPath — Type

GraphIsPath()

Is an undirected graph a path?

julia> using GraphProperties

julia> GraphIsPath() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPerfect — Type

GraphIsPerfect()

Is an undirected graph perfect?

julia> using GraphProperties

julia> GraphIsPerfect() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPermutationGraph — Type

GraphIsPermutationGraph()

Is an undirected graph a permutation graph?

julia> using GraphProperties

julia> GraphIsPermutationGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPlanar — Type

GraphIsPlanar()

Is an undirected graph planar?

julia> using GraphProperties

julia> GraphIsPlanar() isa GraphProperty{Bool}
true

GraphProperties.GraphIsPtolemaic — Type

GraphIsPtolemaic()

Is an undirected graph Ptolemaic?

julia> using GraphProperties

julia> GraphIsPtolemaic() isa GraphProperty{Bool}
true

GraphProperties.GraphIsRegular — Type

GraphIsRegular()

Is an undirected graph regular?

julia> using GraphProperties

julia> GraphIsRegular() isa GraphProperty{Bool}
true

GraphProperties.GraphIsSplitGraph — Type

GraphIsSplitGraph()

Is an undirected graph a split graph?

julia> using GraphProperties

julia> GraphIsSplitGraph() isa GraphProperty{Bool}
true

GraphProperties.GraphIsTree — Type

GraphIsTree()

Is an undirected graph a tree?

julia> using GraphProperties

julia> GraphIsTree() isa GraphProperty{Bool}
true

GraphProperties.GraphIsTriangleFree — Type

GraphIsTriangleFree()

Is an undirected graph triangle-free?

julia> using GraphProperties

julia> GraphIsTriangleFree() isa GraphProperty{Bool}
true

GraphProperties.GraphIsTriviallyPerfect — Type

GraphIsTriviallyPerfect()

Is an undirected graph trivially perfect?

julia> using GraphProperties

julia> GraphIsTriviallyPerfect() isa GraphProperty{Bool}
true

GraphProperties.HadwigerNumber — Type

HadwigerNumber()

The Hadwiger number of an undirected graph. Also known as the contraction clique number or as the homomorphism degree.

julia> using GraphProperties

julia> HadwigerNumber() isa GraphProperty{Integer}
true

GraphProperties.IndependenceNumber — Type

IndependenceNumber()

The independence number of an undirected graph.

julia> using GraphProperties

julia> IndependenceNumber() isa GraphProperty{Integer}
true

GraphProperties.IntersectionNumber — Type

IntersectionNumber()

The intersection number of an undirected graph. Also known as the R-content or as the edge clique cover number or as the clique cover number.

julia> using GraphProperties

julia> IntersectionNumber() isa GraphProperty{Integer}
true

GraphProperties.IsDirectedGraph — Type

IsDirectedGraph()

Is something a directed graph?

julia> using GraphProperties

julia> IsDirectedGraph() isa GraphProperty{Bool}
true

GraphProperties.IsUndirectedGraph — Type

IsUndirectedGraph()

Is something an undirected graph?

julia> using GraphProperties

julia> IsUndirectedGraph() isa GraphProperty{Bool}
true

GraphProperties.MatchingNumber — Type

MatchingNumber()

The matching number of an undirected graph.

julia> using GraphProperties

julia> MatchingNumber() isa GraphProperty{Integer}
true

GraphProperties.MaximumDegree — Type

MaximumDegree()

The maximum degree among the degrees of the vertices of an undirected graph.

julia> using GraphProperties

julia> MaximumDegree() isa GraphProperty{Integer}
true

GraphProperties.MinimumDegree — Type

MinimumDegree()

The minimum degree among the degrees of the vertices of an undirected graph.

julia> using GraphProperties

julia> MinimumDegree() isa GraphProperty{Integer}
true

GraphProperties.NumberOfArcs — Type

NumberOfArcs()

The number of arcs of a directed graph.

julia> using GraphProperties

julia> NumberOfArcs() isa GraphProperty{Integer}
true

GraphProperties.NumberOfConnectedComponents — Type

NumberOfConnectedComponents()

The number of connected components of an undirected graph.

julia> using GraphProperties

julia> NumberOfConnectedComponents() isa GraphProperty{Integer}
true

GraphProperties.NumberOfEdges — Type

NumberOfEdges()

The number of edges of an undirected graph.

julia> using GraphProperties

julia> NumberOfEdges() isa GraphProperty{Integer}
true

GraphProperties.NumberOfVertices — Type

NumberOfVertices()

The number of vertices of a graph.

julia> using GraphProperties

julia> NumberOfVertices() isa GraphProperty{Integer}
true

GraphProperties.Pathwidth — Type

Pathwidth()

The pathwidth of an undirected graph. Also known as the interval thickness or as the vertex separation number or as the node searching number.

julia> using GraphProperties

julia> Pathwidth() isa GraphProperty{Integer}
true

GraphProperties.Sphericity — Type

Sphericity()

The sphericity of an undirected graph.

julia> using GraphProperties

julia> Sphericity() isa GraphProperty{Integer}
true

GraphProperties.Splittance — Type

Splittance()

The splittance of an undirected graph.

julia> using GraphProperties

julia> Splittance() isa GraphProperty{Integer}
true

GraphProperties.Strength — Type

Strength()

The strength of an undirected graph.

julia> using GraphProperties

julia> Strength() isa GraphProperty{Real}
true

GraphProperties.StrongDominationNumber — Type

StrongDominationNumber()

The strong domination number of an undirected graph.

julia> using GraphProperties

julia> StrongDominationNumber() isa GraphProperty{Integer}
true

GraphProperties.Treewidth — Type

Treewidth()

The treewidth of an undirected graph.

julia> using GraphProperties

julia> Treewidth() isa GraphProperty{Integer}
true

GraphProperties.TwinWidth — Type

TwinWidth()

The twin-width of an undirected graph.

julia> using GraphProperties

julia> TwinWidth() isa GraphProperty{Integer}
true

GraphProperties.VertexConnectivity — Type

VertexConnectivity()

The vertex-connectivity of an undirected graph. Also known as the connectivity.

julia> using GraphProperties

julia> VertexConnectivity() isa GraphProperty{Integer}
true

GraphProperties.VertexCoverNumber — Type

VertexCoverNumber()

The vertex cover number of an undirected graph.

julia> using GraphProperties

julia> VertexCoverNumber() isa GraphProperty{Integer}
true